+++ services/director/src/simcore_service_director/modules/docker_registry/__init__.py
@@ -1,0 +1,7 @@
+from . import _client as client
+from ._setup import setup
+
+__all__: tuple[str, ...] = (
+    "client",
+    "setup",
+)
